+++ PliP/src/com/plip/persistence/dao/impls/ProductDaoImpl.java
@@ -13,22 +13,19 @@ import com.plip.persistence.model.Status;
 
 public class ProductDaoImpl implements ProductDao{
 
-	
-	
 	public ProductDaoImpl() {
-		super();
-		
+		super();		
 	}
 
 	@Override
-	public Integer addProduct(Product product) {
+	public Long addProduct(Product product) {
 		SessionFactory factory = DaoManager.createSessionFactory();
 		Session session = factory.openSession();
 		Transaction tx = null;
-		Integer productID = null;
+		Long productID = null;
 		try {
 			tx = session.beginTransaction();
-			productID = (Integer) session.save(product);
+			productID = (Long) session.save(product);
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
