+++ PliP/src/com/plip/persistence/dao/interfaces/ProductDao.java
@@ -4,7 +4,7 @@ import com.plip.persistence.model.Product;
 
 public interface ProductDao {
 	
-	public Integer addProduct(Product product);
+	public Long addProduct(Product product);
 	
 	public Product getProduct(int idProduct);
